requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
rank-bm25
pydantic
python-dotenv
tiktoken
openai
numpy
pandas
loguru
pypdf
python-docx
langchain
langchain-text-splitters